harden FEC wire/manifest validation + enable overflow checks
- framing: reject FEC shard_count > 256 (the galois_8 RS limit) before the
  shard-metadata allocations, closing a 32-bit usize multiply-overflow that
  could under-allocate and overrun on read.
- receiver: validate manifest FecParams (data/parity ≥ 1, sum ≤ 256) up front
  with a clear error instead of failing deep in stripe-buffer/RS init. Adds
  unit tests for the new checks.
- Cargo: enable overflow-checks in the release profile — wire-driven size math
  should trap, not wrap.
- negotiate: drop two u64->u64 casts that were breaking `clippy -D warnings`
  (and thus CI).

+++ src/transfer/negotiate.rs
@@ -411,7 +411,7 @@ mod tests {
             );
             // Stagger timestamps so duration is meaningful.
             s.timestamp = std::time::Instant::now()
-                .checked_sub(std::time::Duration::from_millis((400 - i * 100) as u64))
+                .checked_sub(std::time::Duration::from_millis(400 - i * 100))
                 .unwrap_or_else(std::time::Instant::now);
             samples.push(s);
         }
@@ -438,7 +438,7 @@ mod tests {
         for i in 0..5u64 {
             let mut s = make_sample(5_000_000 + i * 1_000, 1, 0);
             s.timestamp = std::time::Instant::now()
-                .checked_sub(std::time::Duration::from_millis((400 - i * 100) as u64))
+                .checked_sub(std::time::Duration::from_millis(400 - i * 100))
                 .unwrap_or_else(std::time::Instant::now);
             samples.push(s);
         }
